chore(client): Phase 3 Step 21 – migrate Register.jsx to react-hook-form + zod; unify validation and loading states\nchore(mobile): Phase 4 Step 36 – move providers to shared/providers with re-export stubs; update providers barrel\ndocs: update URBAN_REALTY_REFACTORING_MASTER_REPORT and Phase 3/4 logs
