+++ mobile/lib/providers/theme_provider.dart
@@ -1,87 +1,1 @@
-import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-class ThemeProvider extends ChangeNotifier {
-  static const String _themeKey = 'theme_mode';
-  late ThemeMode _themeMode;
-  
-  ThemeProvider() {
-    _themeMode = ThemeMode.system;
-    _loadThemeMode();
-  }
-  
-  ThemeMode get themeMode => _themeMode;
-  
-  bool get isDarkMode => _themeMode == ThemeMode.dark;
-  bool get isLightMode => _themeMode == ThemeMode.light;
-  bool get isSystemMode => _themeMode == ThemeMode.system;
-  
-  Future<void> _loadThemeMode() async {
-    try {
-      final prefs = await SharedPreferences.getInstance();
-      final themeString = prefs.getString(_themeKey);
-      if (themeString != null) {
-        switch (themeString) {
-          case 'light':
-            _themeMode = ThemeMode.light;
-            break;
-          case 'dark':
-            _themeMode = ThemeMode.dark;
-            break;
-          case 'system':
-            _themeMode = ThemeMode.system;
-            break;
-        }
-        notifyListeners();
-      }
-    } catch (e) {
-      // Keep default theme mode if loading fails
-      debugPrint('Failed to load theme mode: $e');
-    }
-  }
-  
-  Future<void> setThemeMode(ThemeMode mode) async {
-    if (_themeMode == mode) return;
-    
-    _themeMode = mode;
-    notifyListeners();
-    
-    try {
-      final prefs = await SharedPreferences.getInstance();
-      String themeString;
-      switch (mode) {
-        case ThemeMode.light:
-          themeString = 'light';
-          break;
-        case ThemeMode.dark:
-          themeString = 'dark';
-          break;
-        case ThemeMode.system:
-          themeString = 'system';
-          break;
-      }
-      await prefs.setString(_themeKey, themeString);
-    } catch (e) {
-      debugPrint('Failed to save theme mode: $e');
-    }
-  }
-  
-  Future<void> toggleTheme() async {
-    if (_themeMode == ThemeMode.light) {
-      await setThemeMode(ThemeMode.dark);
-    } else {
-      await setThemeMode(ThemeMode.light);
-    }
-  }
-  
-  String getThemeModeString() {
-    switch (_themeMode) {
-      case ThemeMode.light:
-        return 'Light';
-      case ThemeMode.dark:
-        return 'Dark';
-      case ThemeMode.system:
-        return 'System';
-    }
-  }
-}
+export '../shared/providers/theme_provider.dart';
