+++ mobile/lib/shared/providers/index.dart
@@ -1,6 +1,3 @@
-export '../../providers/auth_provider.dart';
-export '../../providers/properties_provider.dart';
-export '../../providers/theme_provider.dart';
-export * from "./auth_provider.dart";
-export * from "./properties_provider.dart";
-export * from "./theme_provider.dart";
+export * from './auth_provider.dart';
+export * from './properties_provider.dart';
+export * from './theme_provider.dart';
